Refactor: useProfileActions hook 생성 및 적용 #45
